Converted dust type 0 to 60fps and fixed two bugs

+++ src/main/java/legend/game/submap/DustType0Particle20.java
@@ -7,7 +7,7 @@ public class DustType0Particle20 {
   public float stepSize_04;
   public float size_08;
   public final Vector3f transfer = new Vector3f();
-  public short maxTicks_18;
+  public int maxTicks_18;
 
   public DustType0Particle20 next_1c;
 }
